Final build

+++ Assets/Scenes/Dev_Scene/Nicolas/Script/DataScriptableObject/ClassData.cs
@@ -6,9 +6,9 @@ public class ClassData : ScriptableObject
     public string description;
     public GameObject gameObjectmodel;
 
-    public int classArmor;
-    public int classSpeed;
-    public int classDamage;
+    public float classArmor;
+    public float classSpeed;
+    public float classDamage;
     public float cooldownAttack;
     public float cooldownSkill;
